Change the behavior of global search:
Add an option Feature ID Match that looks for strings in the query that match feature IDs, and
drive the browser directly to either the feature page, if a single feature was selected,
or to the feature list page configured with the set of features.

Add an option Genome ID Match that looks for strings in the query that match genome IDs, and
drive the browser directly to either the genome page, if a single genome was selected,
or to the genome list page configured with the set of genomes.

If neither of these options was selected, but the query text includes either a feature ID or a
genome ID, act as if the Feature ID Match or Genome ID Match option was selected.