Reduce facet overhead on filter bar
If getFacets() is invoked without a facetFields parameter, return a failed promise.  This case is invoked during the filter panel initialization and results in the creation of a Solr query with potentially a very large number of facets and is quite expensive. 

This change does not appear to change the behavior of the tool.